requirements.txt
fastapi
uvicorn
gtts
requests
python-dotenv
adobe-pdfservices-sdk 
pydantic
transformers 
torch
cleantext
gtts